requirements.txt
pandas
numpy
Flask
Pillow
tensorflow
-e .